Disable IPv6 temp addressing and enable smartd mons
Keeping IPv6 privacy extensions enabled leads to rotating IPv6
addresses. That is not ideal on a server environment.

Moreover, the DDNS service fails to determine the GUA and instead
incorrectly uses the temporary GUA, which leads to failure in routing.

This patch also enables smartd monitoring and notifications using msmtp.

Signed-off-by: Nick <nickk.2974@gmail.com>